requirements.txt
fastmcp
supabase
python-dotenv
pg8000
